requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Authentication (Firebase)
firebase-admin==6.3.0

# Database and Storage
sqlalchemy==2.0.23
chromadb==0.4.18
google-cloud-storage==2.13.0

# AI/ML
openai==1.3.7
pillow==10.1.0

# LangChain and LangGraph (compatible versions)
langchain==0.3.17
langgraph==0.2.70
langchain-openai==0.3.0
langchain-community==0.3.17
langchain-core==0.3.28

# Embeddings
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.1

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.1
ruff==0.1.6
